screentime function functional

+++ src/main/java/com/example/digitaldetox/tracker/App.java
@@ -3,11 +3,11 @@ package com.example.digitaldetox.tracker;
 public class App {
 
     private String name;
-    private int app_time;
+    private long app_time;
     private int overall_time;
     private int accountID;
 
-    public App(int accountID, String name, int app_time) {
+    public App(int accountID, String name, long app_time) {
         this.name = name;
         this.app_time = app_time;
         this.accountID = accountID;
@@ -22,7 +22,7 @@ public class App {
         return name;
     }
 
-    public int getTime() { return app_time; }
+    public long getTime() { return app_time; }
 
     public void setName(String name) {
         this.name = name;
